#290 - rename test classes
